pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui_snacknodes"
description = "Integration of functional ComfyUI nodes. Designed to reduce reliance on multiple third-party node packages by providing a single comprehensive solution for all needs."
version = "0.0.3"
license = {text = "GNU General Public License v3"}
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0"
]

[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
bugs = "https://github.com/linksluckytime/comfyui_snacknodes/issues"
homepage = "https://github.com/linksluckytime/comfyui_snacknodes"

[tool.comfy]
PublisherId = "linksluckytime"
DisplayName = "comfyui_snacknodes"
Icon = "https://raw.githubusercontent.com/linksluckytime/comfyui_snacknodes/main/web/icon.png"

[tool.setuptools]
packages = ["comfyui_snacknodes", "comfyui_snacknodes.nodes", "comfyui_snacknodes.utils", "comfyui_snacknodes.config"]
package-dir = {"comfyui_snacknodes" = "."}

[tool.setuptools.package-data]
"*" = ["*.png"]
"comfyui_snacknodes" = ["web/*"]